+++ DungeonBuilder/dungeon_serializer.cpp
@@ -17,7 +17,7 @@ extern GlobalState globalState;
 int _loadInt(string name,void *_json)
 {
 	json_value * json = (json_value*)_json;
-	for(int i = 0; i < json->u.object.length; i++)
+	for(size_t i = 0; i < json->u.object.length; i++)
 	{
 		string json_name = json->u.object.values[i].name;
 		if(json_name == name)
@@ -34,7 +34,7 @@ int _loadInt(string name,void *_json)
 string _loadString(string name,void *_json)
 {
 	json_value * json = (json_value*)_json;
-	for(int i = 0; i < json->u.object.length; i++)
+	for(size_t i = 0; i < json->u.object.length; i++)
 	{
 		string json_name = json->u.object.values[i].name;
 		if(json_name == name)
@@ -51,7 +51,7 @@ string _loadString(string name,void *_json)
 bool _loadBool(string name,void *_json)
 {
 	json_value * json = (json_value*)_json;
-	for(int i = 0; i < json->u.object.length; i++)
+	for(size_t i = 0; i < json->u.object.length; i++)
 	{
 		string json_name = json->u.object.values[i].name;
 		if(json_name == name)
@@ -69,7 +69,7 @@ vector<string> _loadVectorString(string name,void *_json)
 {
 	json_value * json = (json_value*)_json;
 	vector<string> result;
-	for(int i = 0; i < json->u.object.length; i++)
+	for(size_t i = 0; i < json->u.object.length; i++)
 	{
 		string json_name = json->u.object.values[i].name;
 		if(json_name == name)
@@ -77,7 +77,7 @@ vector<string> _loadVectorString(string name,void *_json)
 			if(json->u.object.values[i].value->type == json_array)
 			{
 				auto v = json->u.object.values[i].value->u.array;
-				for(int j = 0; j < v.length; j++)
+				for(size_t j = 0; j < v.length; j++)
 				{
 					if(v.values[j]->type == json_string) {
 						string s = v.values[j]->u.string.ptr;
@@ -99,7 +99,7 @@ vector<string> _loadVectorString(string name,void *_json)
 void _loadEntity(string name,DungeonEntity **e,void *_json)
 {
 	json_value * json = (json_value*)_json;
-	for(int i = 0; i < json->u.object.length; i++)
+	for(size_t i = 0; i < json->u.object.length; i++)
 	{
 		string json_name = json->u.object.values[i].name;
 		if(json_name == name)
@@ -118,7 +118,7 @@ void _loadVectorEntity(string name,void *_v,void *_json)
 {
 	json_value * json = (json_value*)_json;
 	vector<DungeonEntity*> *result = (vector<DungeonEntity*> *)_v;
-	for(int i = 0; i < json->u.object.length; i++)
+	for(size_t i = 0; i < json->u.object.length; i++)
 	{
 		string json_name = json->u.object.values[i].name;
 		if(json_name == name)
@@ -126,7 +126,7 @@ void _loadVectorEntity(string name,void *_v,void *_json)
 			if(json->u.object.values[i].value->type == json_array)
 			{
 				auto v = json->u.object.values[i].value->u.array;
-				for(int j = 0; j < v.length; j++)
+				for(size_t j = 0; j < v.length; j++)
 				{
 					if(v.values[j]->type == json_integer) {
 						int uid = v.values[j]->u.integer;
@@ -169,7 +169,7 @@ string _writeBool(string name,bool value)
 string _writeVectorString(string name,vector<string> const &value)
 {
 	string result =  STR_QUOT + name + STR_QUOT+ ":[";
-	for(unsigned int i = 0; i < value.size(); i++)
+	for(size_t i = 0; i < value.size(); i++)
 	{
 		result += STR_QUOT + value[i] + STR_QUOT;
 		if(i != value.size()-1) {
@@ -217,14 +217,14 @@ void loadJson(string filename)
 
 	//shallow load of all entities
 	json_value* root = json_parse(jsonStr,contents.size()+1);
-	for(int i = 0; i < root->u.object.length;i++)
+	for(size_t i = 0; i < root->u.object.length;i++)
 	{
 		auto v = root->u.object.values[i].value->u.array;
 		string name = root->u.object.values[i].name;
 		if(name == "rooms")
 		{
 
-			for(int j = 0; j < v.length; j++)
+			for(size_t j = 0; j < v.length; j++)
 			{
 				DungeonRoom *room = new DungeonRoom(v.values[j]);
 				if(j == 0) globalState.startRoom = room;				
@@ -233,42 +233,42 @@ void loadJson(string filename)
 		}
 		else if(name == "exits")
 		{
-			for(int j = 0; j < v.length; j++)
+			for(size_t j = 0; j < v.length; j++)
 			{
 				DungeonExit *exit = new DungeonExit(v.values[j]);								
 			}
 		}
 		else if(name == "creatures")
 		{
-			for(int j = 0; j < v.length; j++)
+			for(size_t j = 0; j < v.length; j++)
 			{
 				DungeonCreature *creature = new DungeonCreature(v.values[j]);								
 			}
 		}
 		else if(name == "objects")
 		{
-			for(int j = 0; j < v.length; j++)
+			for(size_t j = 0; j < v.length; j++)
 			{
 				DungeonObject *object = new DungeonObject(v.values[j]);								
 			}
 		}
 		else if(name == "effects")
 		{
-			for(int j = 0; j < v.length; j++)
+			for(size_t j = 0; j < v.length; j++)
 			{
 				DungeonEffect *effect = new DungeonEffect(v.values[j]);								
 			}
 		}
 		else if(name == "actions")
 		{
-			for(int j = 0; j < v.length; j++)
+			for(size_t j = 0; j < v.length; j++)
 			{
 				DungeonAction *action = new DungeonAction(v.values[j]);								
 			}
 		}
 		else if(name == "triggers")
 		{
-			for(int j = 0; j < v.length; j++)
+			for(size_t j = 0; j < v.length; j++)
 			{
 				DungeonTrigger *trigger = new DungeonTrigger(v.values[j]);							
 			}
